made page validate
woohoo! reveal.js is now HTML5 compliant! 

this is validated using the http://validator.w3.org/ and the http://html5.validator.nu/ services.

this was accomplished using a tad of JavaScript to get around a fullscreen bug.

This also (slightly unintentionally but helpfully) solves a tiny bug with fullscreening in old Opera 11.52 (I think, could be 11.51 or other)